+++ crates/miden-testing/src/kernel_tests/tx/test_fpi.rs
@@ -19,7 +19,6 @@ use miden_protocol::account::{
 use miden_protocol::assembly::DefaultSourceManager;
 use miden_protocol::asset::{
     Asset,
-    AssetCallbackFlag,
     AssetVaultKey,
     FungibleAsset,
     NonFungibleAsset,
@@ -757,8 +756,7 @@ async fn foreign_account_can_get_balance_and_presence_of_asset() -> anyhow::Resu
     let non_fungible_asset = Asset::NonFungible(NonFungibleAsset::new(
         &NonFungibleAssetDetails::new(non_fungible_faucet_id, vec![1, 2, 3]),
     ));
-    let fungible_asset_key =
-        AssetVaultKey::new_fungible(fungible_faucet_id, AssetCallbackFlag::Disabled);
+    let fungible_asset_key = AssetVaultKey::new_fungible(fungible_faucet_id);
 
     let foreign_account_code_source = format!(
         "
@@ -872,8 +870,7 @@ async fn foreign_account_can_get_balance_and_presence_of_asset() -> anyhow::Resu
 async fn foreign_account_get_initial_balance() -> anyhow::Result<()> {
     let fungible_faucet_id = AccountId::try_from(ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_1)?;
     let fungible_asset = Asset::Fungible(FungibleAsset::new(fungible_faucet_id, 10)?);
-    let fungible_asset_key =
-        AssetVaultKey::new_fungible(fungible_faucet_id, AssetCallbackFlag::Disabled);
+    let fungible_asset_key = AssetVaultKey::new_fungible(fungible_faucet_id);
 
     let foreign_account_code_source = format!(
         "